% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_char_vectors.r
\name{compare_char_vectors}
\alias{compare_char_vectors}
\alias{cosine_similarity}
\title{Compare the similarity of two vectors}
\usage{
compare_char_vectors(lhsv, rhsv)

cosine_similarity(xvector, yvector)
}
\arguments{
\item{lhsv}{left handside character vector}

\item{rhsv}{right handside character vector}

\item{xvector}{numeric vector of x}

\item{yvector}{numberic vector of y}
}
\value{
`compare_char_vector()` return a one number score between 0~1 indicating
the cosine similarity of the two input vectors

`cosine_similarity()` return a one number score between 0~1 indicating
the cosine similarity of the two input vectors
}
\description{
This is the fundamental function of comparing two vectors
and return cosine similarity score, suggest to use with *apply funciton
}
\details{
The output of `compare_char_vectors()` will be a one number score between 0~1

The output of `cosine_similarity()` will be a one number score between 0~1
}
\examples{
compare_char_vectors(c('one','two','three'),c('two','three','four'))
cosine_similarity(c(1,2,3),c(2,3,4))
}
