% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2df.r
\name{json2df}
\alias{json2df}
\title{convert json data to data.frame}
\usage{
json2df(jsonfilename)
}
\arguments{
\item{jsonfilename}{path of json file}
}
\value{
`json2df()` return a data.frame obj
}
\description{
convert json data to data.frame
}
\examples{
json2df("./extensions.json")
}
