% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit_index.r
\name{model_fit_index}
\alias{model_fit_index}
\title{Calculate model fitting index}
\usage{
model_fit_index(y, yhat, nbeta)
}
\arguments{
\item{y}{original target value vector}

\item{yhat}{predicted target value vector}

\item{nbeta}{number of beta estimated(including constant)}
}
\value{
`model_fit_index()` return a list of all possible normally used index of evaluating a model
}
\description{
This will a function of computing modeling index, return such as r-square, rmse, etc.
}
\details{
The output of `model_fit_index()` will be a one list of all indexes
}
\examples{
model_fit_index(y,yhat)
}
